Following changes: - Transition revision because of midside node detection during inner 2D-water level boundary condition calculation - Transition revision because drying section used wrong parameter - some redeclarations - deactivation of autoconverge code; it doesn't work yet

+++ _import/branches/rma10sk_branches/updateSubroutines/RMA10S_TUHH/source/QKalypsoCalc.F90
@@ -21,7 +21,7 @@ INTEGER               :: na, lump
 INTEGER               :: k
 !geometric waterdepth, Marsh-waterdepth
 REAL (KIND=8)         :: fliesstiefe, d3
-real (KIND=4)         :: d2
+real (KIND=8)         :: d2
 !new variables
 REAL                  :: amec (350)
